+++ src/main/java/com/greedy/zupzup/member/domain/Member.java
@@ -1,4 +1,4 @@
-package com.greedy.zupzup.member;
+package com.greedy.zupzup.member.domain;
 
 import com.greedy.zupzup.global.BaseTimeEntity;
 import lombok.*;
