+++ src/main/java/com/greedy/zupzup/member/domain/Provider.java
@@ -1,4 +1,4 @@
-package com.greedy.zupzup.member;
+package com.greedy.zupzup.member.domain;
 
 public enum Provider {
     GOOGLE,
